Enable optimize_for_size for 3P
Turn on the optimize_for_size wich for the most
part just enables -Os instead of -O2. The flag is on
for Android but for some reason was not turned of for 3P.
This optimization was also used in Cobalt 25.

The result is 14MB binary size reduction and ~ 10MB memory reduction.

 libcobalt.so 85MB -> 71MB
 libcobalt.lz4 52MB -> 43MB
 RSS home, signed out 135MB -> 124MB

Issue: 421171908